git-svn-id: svn+ssh://statnet.csde.washington.edu/svn/statnet_commons/ergm/trunk@14627 9cb281f4-553f-406f-8d7f-11f072bbfe46

+++ src/CD.c
@@ -259,7 +259,7 @@ MCMCStatus CDStep(MHproposal *MHp,
     }
     
     /* if we accept the proposed network */
-    if (cutoff >= 0.0 || log(unif_rand()) < cutoff) { 
+    if (cutoff >= 0.0 || logf(unif_rand()) < cutoff) { 
       if(fVerbose>=5){
 	Rprintf("Accepted.\n");
       }
